refactor(): Packaging
Adds new packaging schema and refactors import in a few classes
