+++ EmployeeWorkOrderManagment/src/main/java/ls/EmployeeWorkOrderManagment/web/controller/HomepageController.java
@@ -1,4 +1,4 @@
-package ls.EmployeeWorkOrderManagment.controller;
+package ls.EmployeeWorkOrderManagment.web.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
